requirements.txt
transformers==4.41.2
pillow==10.3.0
chardet==5.2.0
flash_attn==2.5.8
accelerate==0.30.1
bitsandbytes==0.43.1
peft==0.11.1
timm==1.0.3
sentencepiece==0.2.0
torchvision~=0.18.0
trl~=0.8.6
thefuzz==0.22.1
loguru~=0.7.2